implemented classifier optimisation script, among other changes
additionally fixed incorrect computation of TPR and FPR for region detection models, as well as fixed the dataloaders to be compatible with concatenated datasets. Also fixed MIAS and DDSM datasets loading incorrectly with the dataloader classes. Various other minor bug fixes are included, especially regarding correcting the dimensionalities of the default classification model.